feat: 쿠폰 컨트롤러에 서비스 적용 - 보유 쿠폰 목록 조회 기능
