+++ src/main/java/kr/hhplus/be/server/interfaces/coupon/CouponController.java
@@ -2,8 +2,7 @@ package kr.hhplus.be.server.interfaces.coupon;
 
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Positive;
-import kr.hhplus.be.server.interfaces.coupon.dto.CouponIssueRequest;
-import kr.hhplus.be.server.interfaces.coupon.dto.CouponResponse;
+import kr.hhplus.be.server.domain.coupon.CouponService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -14,23 +13,33 @@ import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
 
+import static kr.hhplus.be.server.interfaces.coupon.CouponRequest.*;
+import static kr.hhplus.be.server.interfaces.coupon.CouponResponse.*;
+
 @RestController
 @RequestMapping("/api/v1/coupons")
 public class CouponController implements CouponApiSpec {
 
+    private final CouponService couponService;
+
+    public CouponController(CouponService couponService) {
+        this.couponService = couponService;
+    }
+
     @GetMapping
     @Override
-    public ResponseEntity<List<CouponResponse>> getUserCoupons(@RequestParam @Positive long userId) {
-        return ResponseEntity.ok(List.of(
-                new CouponResponse(1L, "쿠폰1", "2025-12-31", false),
-                new CouponResponse(2L, "쿠폰2", "2025-12-31", true),
-                new CouponResponse(3L, "쿠폰3", "2025-05-31", false)
-        ));
+    public ResponseEntity<List<UserCouponResponse>> getUserCoupons(@RequestParam @Positive long userId) {
+
+        List<UserCouponResponse> response = couponService.findByUserId(userId).stream()
+                .map(UserCouponResponse::from)
+                .toList();
+
+        return ResponseEntity.ok(response);
     }
 
     @PostMapping
     @Override
-    public ResponseEntity<CouponResponse> issueCoupon(@RequestBody @Valid CouponIssueRequest request) {
-        return ResponseEntity.ok(new CouponResponse(1L, "쿠폰1", "2025-12-31", false));
+    public ResponseEntity<UserCouponResponse> issueCoupon(@RequestBody @Valid CouponIssueRequest request) {
+        return ResponseEntity.ok(new UserCouponResponse(1L, "쿠폰1", "2025-12-31", false));
     }
 }
